Add delivery_carrier_pricelist
New module allowing to compute a delivery cost based on the sales order
pricelist.

Most of the code of the module is to update the many 'attrs' of the
'delivery' module which have domains based on the "delivery_type" field
and cannot be extended in XML without breaking compatibility.

When using an external provider (such as DHL, UPS), the "Pricelist"
provider cannot be used. In this case, the invoice policy, which is by
default "Estimate" or "Real" has a third option "Pricelist Cost". This
option would not make sense with "Estimate" or "Real", which is why this
field is used.